Move startTls and `Location`
