+++ cli/js/deno_unstable.ts
@@ -15,3 +15,4 @@ export { utimeSync, utime } from "./ops/fs/utime.ts";
 export { ShutdownMode, shutdown } from "./net.ts";
 export { listen, listenDatagram, connect } from "./net_unstable.ts";
 export { cwd } from "./ops/fs/dir.ts";
+export { startTls } from "./tls.ts";
