+++ cli/js/lib.deno.unstable.d.ts
@@ -664,6 +664,17 @@ declare namespace Deno {
     options?: CompilerOptions
   ): Promise<[DiagnosticItem[] | undefined, string]>;
 
+  /** **UNSTABLE**: Should not have same name as `window.location` type. */
+  interface Location {
+    /** The full url for the module, e.g. `file://some/file.ts` or
+     * `https://some/file.ts`. */
+    fileName: string;
+    /** The line number in the file. It is assumed to be 1-indexed. */
+    lineNumber: number;
+    /** The column number in the file. It is assumed to be 1-indexed. */
+    columnNumber: number;
+  }
+
   /** UNSTABLE: new API, yet to be vetted.
    *
    * Given a current location in a module, lookup the source location and return
@@ -955,4 +966,31 @@ declare namespace Deno {
    * Throws `Deno.errors.NotFound` if directory not available.
    */
   export function cwd(): string;
+
+  export interface StartTlsOptions {
+    /** A literal IP address or host name that can be resolved to an IP address.
+     * If not specified, defaults to `127.0.0.1`. */
+    hostname?: string;
+    /** Server certificate file. */
+    certFile?: string;
+  }
+
+  /** **UNSTABLE**: new API, yet to be vetted.
+   *
+   * Start TLS handshake from an existing connection using
+   * an optional cert file, hostname (default is "127.0.0.1").  The
+   * cert file is optional and if not included Mozilla's root certificates will
+   * be used (see also https://github.com/ctz/webpki-roots for specifics)
+   * Using this function requires that the other end of the connection is
+   * prepared for TLS handshake.
+   *
+   *     const conn = await Deno.connect({ port: 80, hostname: "127.0.0.1" });
+   *     const tlsConn = await Deno.startTls(conn, { certFile: "./certs/my_custom_root_CA.pem", hostname: "127.0.0.1", port: 80 });
+   *
+   * Requires `allow-net` permission.
+   */
+  export function startTls(
+    conn: Conn,
+    options?: StartTlsOptions
+  ): Promise<Conn>;
 }
